+++ packages/effects/source/fail.test.ts
@@ -1,16 +1,17 @@
 import { describe, given, it } from '@typed/test'
-import { Capabilities, Yield } from './Effect'
+import { catchFailure, fail } from './fail'
 import { get } from './get'
 import { runEffects } from './runEffects'
-import { fail, handleError } from './throwError'
 
-export const test = describe(`handleError`, [
+export const test = describe(`handleFailure`, [
   given(`an effect that might throw an error`, [
-    it(`allows capturing the error`, ({ equal }) => {
+    it(`allows capturing the error`, ({ equal, same }) => {
       const error = new Error(`Wtf`)
       const fallback = 7
 
       function* throws() {
+        yield* get<{ b: string }>()
+
         yield* fail(error)
       }
 
@@ -23,12 +24,16 @@ export const test = describe(`handleError`, [
       }
 
       function* sut() {
-        const actual = yield* handleError(levelOfIndirection(), () => fallback)
+        const actual = yield* catchFailure(levelOfIndirection(), err => {
+          same(error, err)
+
+          return fallback
+        })
 
         equal(fallback, actual)
       }
 
-      runEffects(sut(), { a: 1 })
+      runEffects(sut(), { a: 1, b: 'asdf' })
     }),
   ]),
 ])
