+++ packages/effects/source/runEffect.ts
@@ -6,11 +6,11 @@ import { startEffect } from './startEffect'
 
 export const runEffect = <A extends Effect<any, any>>(effect: A): Env<Capabilities<A>, Return<A>> =>
   chain(
-    iterator => runEffectIterator<A>(iterator, iterator.next() as IteratorResultOf<A>),
+    iterator => runEffectGenerator<A>(iterator, iterator.next() as IteratorResultOf<A>),
     startEffect<A>(effect),
   )
 
-const runEffectIterator = <A extends Effect<any, any>>(
+const runEffectGenerator = <A extends Effect<any, any>>(
   generator: A,
   result: IteratorResultOf<A>,
 ): Env<Capabilities<A>, Return<A>> =>
@@ -18,7 +18,7 @@ const runEffectIterator = <A extends Effect<any, any>>(
     ? (_: Capabilities<A>) => Resume.of(result.value)
     : chain(
         value =>
-          runEffectIterator<A>(
+          runEffectGenerator<A>(
             generator,
             (value instanceof Failure
               ? generator.return(value.value)
